Fix Milestone E observation routing and hard allNotesOff.
Keep A/B Sensei observation kinds intact, clear voices immediately on
instrument switch/stop, and tighten instrument rack silence tests.

+++ sensei-studio/libs/sensei_engine/include/sensei/engine/instruments/BrightPluck.hpp
@@ -36,7 +36,7 @@ public:
                 break;
             }
         if (voice == nullptr)
-            voice = &voices_[stealIndex_++ % kMaxVoices];
+            voice = &voices_[static_cast<std::size_t>(stealIndex_++ % kMaxVoices)];
 
         voice->active = true;
         voice->midi = midiNote;
